Refactored Quick Sort: renamed file, added dynamic inputs and markdown docs
